renamed css.js to styleFunctionSx.js & fixed lint issues
